Simplify module
