+++ dpc/src/errors/mod.rs
@@ -23,6 +23,9 @@ pub use block::*;
 pub mod dpc;
 pub use dpc::*;
 
+pub mod ledger;
+pub use ledger::*;
+
 pub mod posw;
 pub use posw::*;
