+++ dpc/src/traits/ledger.rs
@@ -14,10 +14,38 @@
 // You should have received a copy of the GNU General Public License
 // along with the snarkVM library. If not, see <https://www.gnu.org/licenses/>.
 
-use crate::Network;
+use crate::{BlockScheme, Network};
 use snarkvm_algorithms::merkle_tree::MerklePath;
 
 use anyhow::Result;
+use std::path::Path;
+
+pub trait LedgerScheme<N: Network>: Sized {
+    type Block: BlockScheme;
+
+    /// Instantiates a new ledger with a genesis block.
+    fn new(path: Option<&Path>, genesis_block: Self::Block) -> Result<Self>;
+
+    /// Returns the latest number of blocks in the ledger.
+    /// A block height of 0 indicates the ledger is uninitialized.
+    /// A block height of 1 indicates the ledger is initialized with a genesis block.
+    fn block_height(&self) -> u32;
+
+    /// Returns the latest block in the ledger.
+    fn latest_block(&self) -> Result<Self::Block>;
+
+    /// Returns the block given the block hash.
+    fn get_block(&self, block_hash: &N::BlockHash) -> Result<Self::Block>;
+
+    /// Returns the block hash given a block number.
+    fn get_block_hash(&self, block_number: u32) -> Result<N::BlockHash>;
+
+    /// Returns the block number given a block hash.
+    fn get_block_number(&self, block_hash: &N::BlockHash) -> Result<u32>;
+
+    /// Returns true if the given block hash exists in the ledger.
+    fn contains_block_hash(&self, block_hash: &N::BlockHash) -> bool;
+}
 
 /// The state commitments tree is a core state tree of the ledger.
 pub trait CommitmentsTreeScheme<N: Network>: Sized {
