+++ ledger/src/ledger/tests.rs
@@ -14,7 +14,7 @@
 // You should have received a copy of the GNU General Public License
 // along with the snarkVM library. If not, see <https://www.gnu.org/licenses/>.
 
-use crate::{ledger::*, prelude::*};
+use crate::ledger::*;
 use snarkvm_dpc::{prelude::*, testnet2::Testnet2};
 
 #[test]
