+++ ledger/src/traits.rs
